AS3 Commons Reflection
The XML based reflection used in this library is extreamly hard to
debug and error prone. This is the start of transitioning this library
over to AS3Commons Reflect.

Edited the getByMetadata and getByMetadataArg functions to work,
without this edit nothing seemed to work correctly.

Added the as3commons reflect dependencies